requirements.txt
ipywidgets==7.7.1
matplotlib
numpy
pandas
torch
torchvision==0.13.1
tqdm==4.64.0
nfl_data_py
seaborn
transformers
scattertext
langchain
openai
tiktoken
pinecone-client
gradio
setuptools==45.0.0